### 2019-08-23 #1680 @kyleniemeyer
Hi @AustinNar, could you add an `affiliation` block to your paper's header, and also modify the author field to include this? It should look something like:
```yaml
authors:
  - name: Austin Nar
    orcid: 0000-0002-5319-9805
    affiliation: 1
affiliations:
 - name:  [department, institution, city, state, country]
   index: 1
```
