### 2019-08-23 #1680 @kyleniemeyer
@AustinNar please remove the "Student" part, and add location information like city, state, and country

For example, my affiliation info would be:
> School of Mechanical, Industrial, and Manufacturing Engineering, Oregon State University, Corvallis, OR, USA
