### 2019-08-23 #1680 @AustinNar
@kyleniemeyer Thank you, I've committed/pushed the updated paper.md
