### 2019-09-16 #1680 @trallard
Hi @AustinNar  apologies for the delay, the first thing we have to do is finding a couple of reviewers do you have someone in mind before moving forward?

If not you can check the list of reviewers at [https://bit.ly/joss-reviewers](https://bit.ly/joss-reviewers)
and see if there is anyone that would be suitable
