### 2019-09-16 #1680 @AustinNar
Hi @trallard no worries! Both @zfurnas and @mbod seem like good choices (background in R & Python and text analysis)
