### 2019-09-24 #1680 @trallard
Excellent!

Dear potential reviewers (@alexanderfurnas and @mbod ),
Based on your expertise in R, Python and text analysis, I would like to ask if you can help us out with a submission to JOSS (The Journal of Open Source Software), where I’m an editor.
Title: tabulog: A Language-Agnostic Template System for Parsing Log Files

Summary: Tabulog is a flexible, powerful framework for parsing log files, specifically designed for web logs (such as the access.log files created by Apache), with the final output being in a tabular format.

Parsing logs with Tabulog requires two things: a template, and a list of "parser classes."

Article Proof: https://github.com/openjournals/joss-papers/blob/joss.01680/joss.01680/10.21105.joss.01680.pdf

Submitting author: @AustinNar  Austin NAr

Repository: https://github.com/AustinNar/tabulog

Version: 2018.12
@alexanderfurnas: Alexander, are you able to and interested in taking on this review?

@mbod : Monica, are you able to and interested in taking on this review?

Your prompt response is appreciated. Please see <https://joss.readthedocs.io/en/latest/reviewer_guidelines.html> for more information about reviewer criteria and the process if it is new to you.
