### 2019-09-24 #1680 @alexanderfurnas
I'm happy to be a reviewer. What is the expected turnaround times for the reviews? Thanks for the reviewer guidelines as well, everything is really well documented (which is not a surprised given the journal), but is a treat.
