### 2019-11-27 #1917 @whedon
[REVIEW]: tabulog: A Language-Agnostic Template System for Parsing Log Files
**Submitting author:** @austinnar (<a href="http://orcid.org/0000-0002-5319-9805">Austin Nar</a>)
**Repository:** <a href="https://github.com/AustinNar/tabulog" target ="_blank">https://github.com/AustinNar/tabulog</a>
**Version:** 0.1.1
