### 2020-02-21 #1917 @AustinNar
Hello! Sorry for the long delay before actually beginning work on this.

I have added a CONTRIBUTING.md which is linked from main README. I added
installation instructions and links to R and Python examples both in the
main README and in a (new) docs/README.md file. Let me know if you would
like any of this to be more comprehensive.

Thanks!

On Sun, Feb 16, 2020 at 10:13 AM Daniel S. Katz <notifications@github.com>
wrote:

> great!
>
