### 2021-01-17 #1917 @arfon
:wave: @AustinNar - as part of a review I've been doing this week as EiC of JOSS of the outstanding submissions I've spent some time reviewing this submission. Unfortunately I believe this submission should have been identified as out of scope during the [pre-review](https://github.com/openjournals/joss-reviews/issues/1680) stage of the process. In particular, this submission is out of scope due to it's small size and [limited scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort).

I realize this submission has been under review for quite some time and I apologize that this was not flagged earlier in the process before @trallard and @alexanderfurnas started their reviews here.
